make PreferenceManager to save app modes and data make home to get location from gps or map Based on the user's decision

+++ app/src/main/java/com/m3/weathervue/model/WeatherResponse.kt
@@ -68,7 +68,7 @@ data class HourlyItem(
 	val dt: Long? = null,
 
 	@field:SerializedName("pop")
-	val pop: Int? = null,
+	val pop: Float? = null,
 
 	@field:SerializedName("wind_deg")
 	val windDeg: Int? = null,
